Change CountingSampler to accept float instead of double (#2592)
Motivation:

`CountingSampler` is currently using `double` for probability, while [Brave](https://github.com/openzipkin/brave/blob/master/brave/src/main/java/brave/sampler/CountingSampler.java#L60) (where this class was originally forked from) is using `float`. This causes issues when passing in floats like `0.01f`, which is converted to `double` as `0.009999999776482582`. This causes `percent` to be `0`.

Modification:

- Change `CountingSampler` to accept a `float`.

Result:

When passing in the probability of `0.01f`, the percentage is correctly calculated as `1`.

+++ core/src/main/java/com/linecorp/armeria/common/util/Sampler.java
@@ -49,7 +49,7 @@ public interface Sampler<T> {
      * @param probability the probability expressed as a floating point number
      *                    between {@code 0.0} and {@code 1.0}.
      */
-    static <T> Sampler<T> random(double probability) {
+    static <T> Sampler<T> random(float probability) {
         return CountingSampler.create(probability);
     }
